+++ exec/call_coverage_heatmaps_Rscript.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+#
+# wrapper bash shell script to run Coverage_Heatmaps_07dec2022.Rmd
+#
+# This script should probably be called:
+# call_coverage_heatmaps_Rscript.sh
+#
+# main goals:
+# define proper RHOME
+# define proper .libPaths
+# call default Rscript which will call the proper R defined by RHOME
+
+